Make the following changes:
,

Find this element:

  <Input placeholder="Enter asset cost" type="number" name="cost" value=0 id="«r4»-form-item" aria-describedby="«r4»-form-item-description" aria-invalid=false className="col-span-3">

Relevant files:

    - https://6000-idx-studio-1745313609480.cluster-w5vd22whf5gmav2vgkomwtc4go.cloudworkstations.dev/_next/static/chunks/src_7f751e95._.js
    - https://6000-idx-studio-1745313609480.cluster-w5vd22whf5gmav2vgkomwtc4go.cloudworkstations.dev/_next/static/chunks/node_modules_efa6a250._.js
    - https://6000-idx-studio-1745313609480.cluster-w5vd22whf5gmav2vgkomwtc4go.cloudworkstations.dev/_next/static/chunks/node_modules_efa6a250._.js
    - https://6000-idx-studio-1745313609480.cluster-w5vd22whf5gmav2vgkomwtc4go.cloudworkstations.dev/_next/static/chunks/src_7f751e95._.js
    - https://6000-idx-studio-1745313609480.cluster-w5vd22whf5gmav2vgkomwtc4go.cloudworkstations.dev/_next/static/chunks/node_modules_efa6a250._.js
    - https://6000-idx-studio-1745313609480.cluster-w5vd22whf5gmav2vgkomwtc4go.cloudworkstations.dev/_next/static/chunks/node_modules_efa6a250._.js
    - https://6000-idx-studio-1745313609480.cluster-w5vd22whf5gmav2vgkomwtc4go.cloudworkstations.dev/_next/static/chunks/node_modules_next_dist_compiled_2ce9398a._.js
    - https://6000-idx-studio-1745313609480.cluster-w5vd22whf5gmav2vgkomwtc4go.cloudworkstations.dev/_next/static/chunks/node_modules_next_dist_compiled_2ce9398a._.js

And change it as follows:

allow all numbers and integers

+++ src/app/assets/list/page.tsx
@@ -87,7 +87,7 @@ const assetFormSchema = z.object({
     message: "Brand must be at least 2 characters.",
   }),
   purchaseDate: z.string().optional(),
-  cost: z.number().int().min(0, {
+  cost: z.number().min(0, {
     message: "Cost must be a positive number.",
   }),
   status: z.enum(['Active', 'Inactive', 'Maintenance']),
@@ -293,3 +293,4 @@ export default function AssetsListPage() {
     </div>
   );
 }
+
